+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.1.5-SNAPSHOT (unreleased)
 
+### Added
+- `Project.projectDependency(...)` and `DependencyHandler.project(...)` https://github.com/RBusarow/kotlin-gradle-extensions/pull/58
+
 ## [0.1.4] - 2023-09-09
 
 ### Changed